Split out utils tests
